+++ operation_status_change.go
@@ -2,6 +2,6 @@ package ravendb
 
 // OperationStatusChange describes a change to the operation status. Can be used as DatabaseChange.
 type OperationStatusChange struct {
-	OperationID int
+	OperationID int64
 	State       map[string]interface{}
 }
